using 32 bit floating point numbers

+++ Intro/Variables/main.go
@@ -3,6 +3,6 @@ package main
 import "fmt"
 
 func main() {
-	var pi float64 = 3.14
+	var pi float32 = 3.14
 	fmt.Printf("Value: %.2f\n", pi)
 }
